requirements.txt
# tight-cli
PyYAML
pytest
boto3
botocore
invoke